Lots of media fixes

+++ app/controllers/MediaController.php
@@ -155,16 +155,16 @@ class mediaController extends \BaseController {
 		// format checkboxes for db
 		$data['reps'] = isset($data['reps']) ? 1 : 0;
 
-		include app_path() . '/helpers/processMedia.php';
+		// process file
+		if ($data['media'] != '') {
+			include app_path() . '/helpers/processMedia.php';
 
-		// if file exists, delete it
-		$old_file = $media->url;
-		if (is_file(public_path() . '/uploads/' . $old_file)) {
-			unlink(public_path() . '/uploads/' . $old_file);
+			// if old file exists, delete it
+			$old_file = $media->url;
+			if (is_file(public_path() . '/uploads/' . $old_file)) {
+				unlink(public_path() . '/uploads/' . $old_file);
+			}
 		}
-
-		// ensure that the file doesn't get saved over with an empty value
-		if ($data['media'] == '') $data['url'] = $media->url;
 		
 		// if role is Superadmin, Admin, or Editor, set owner id to 0
 		if (Auth::user()->hasRole(['Superadmin', 'Admin', 'Editor'])) $data['user_id'] = 0;
@@ -202,8 +202,12 @@ class mediaController extends \BaseController {
 			foreach (Input::get('ids') as $id) {
 				// delete media
 				$media = Media::find($id);
-				unlink(public_path() . '/uploads/' . $media->url);
-				unlink(public_path() . '/uploads/' . $media->image_sm);
+				if (file_exists(public_path() . '/uploads/' . $media->url)) {
+					unlink(public_path() . '/uploads/' . $media->url);
+				}
+				if (file_exists(public_path() . '/uploads/' . $media->image_sm)) {
+					unlink(public_path() . '/uploads/' . $media->image_sm);
+				}
 				Media::destroy($id);
 			}
 			Cache::forget('route_'.Str::slug(action('DataOnlyController@getAllMedia')));
